feat: Update WalletConnect screen
